Create mocks for PackageFiles
